feat: Line이 Section을 포함하는 구조로 변경
