+++ src/main/java/subway/domain/line/domain/Direction.java
@@ -1,4 +1,4 @@
-package subway.domain.section.domain;
+package subway.domain.line.domain;
 
 public enum Direction {
