+++ src/main/java/subway/domain/line/domain/Line.java
@@ -1,0 +1,4 @@
+package subway.domain.line.domain;
+
+public class Line {
+}
