--- src/main/java/subway/domain/section/controller/SectionController.java
@@ -1,41 +1,0 @@
-package subway.domain.section.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.domain.section.dto.SectionCreateRequest;
-import subway.domain.section.dto.SectionDeleteRequest;
-import subway.domain.section.dto.SectionResponse;
-import subway.domain.section.entity.SectionEntity;
-import subway.domain.section.service.SectionService;
-import subway.global.common.ResultResponse;
-
-import javax.validation.Valid;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RestController
-@RequestMapping("/sections")
-public class SectionController {
-
-    private final SectionService sectionService;
-
-    public SectionController(final SectionService sectionService) {
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping
-    public ResponseEntity<ResultResponse> createSection(@RequestBody @Valid final SectionCreateRequest sectionCreateRequest) {
-        final List<SectionEntity> sectionEntities = sectionService.createSection(sectionCreateRequest);
-        final List<SectionResponse> sectionResponses = sectionEntities.stream()
-                .map(SectionResponse::of)
-                .collect(Collectors.toList());
-        return ResponseEntity.status(HttpStatus.CREATED).body(ResultResponse.of(HttpStatus.CREATED, sectionResponses));
-    }
-
-    @DeleteMapping
-    public ResponseEntity<ResultResponse> deleteSection(@RequestBody @Valid final SectionDeleteRequest sectionDeleteRequest) {
-        sectionService.deleteSection(sectionDeleteRequest.getLineId(), sectionDeleteRequest.getStationId());
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body(ResultResponse.of(HttpStatus.NO_CONTENT, sectionDeleteRequest.getStationId()));
-    }
-}
