+++ src/test/java/subway/section/dao/RdsSectionDaoTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import subway.domain.section.dao.RdsSectionDao;
-import subway.domain.section.domain.Direction;
-import subway.domain.section.entity.SectionEntity;
+import subway.domain.line.dao.RdsSectionDao;
+import subway.domain.line.domain.Direction;
+import subway.domain.line.entity.SectionEntity;
 
 import javax.sql.DataSource;
 import java.util.List;
