+++ src/test/java/subway/section/dao/StubSectionDao.java
@@ -1,8 +1,8 @@
 package subway.section.dao;
 
-import subway.domain.section.dao.SectionDao;
-import subway.domain.section.domain.Direction;
-import subway.domain.section.entity.SectionEntity;
+import subway.domain.line.dao.SectionDao;
+import subway.domain.line.domain.Direction;
+import subway.domain.line.entity.SectionEntity;
 
 import java.util.HashMap;
 import java.util.List;
